requirements.txt
customtkinter==5.2.2
fpdf2==2.8.7
openpyxl==3.1.5
python-docx==1.2.0
Pillow==12.1.1
PyMuPDF==1.25.3
